requirements.txt
python-telegram-bot==20.7
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0
tinydb==4.8.0
